split out the UpnpClient logic from the NatManager
